ARM: Integrator: pass 'khz' to integrator_time_init
This is now what the clocksource/clockevent initialization functions
want, so give them the timer tick rate directly.

+++ arch/arm/mach-integrator/common.h
@@ -1,2 +1,2 @@
 FASTVC-BENCH-CORPUS:linux-main-100k-v1-e0bd41dc-8e12-4f1b-978d-a3645ae59da0
-extern void integrator_time_init(unsigned long, unsigned int);
+extern void integrator_time_init(u32, unsigned int);
